U-update

+++ install_kernel.sh
@@ -271,14 +271,28 @@ virt_check(){
 
 function installSoftDownload(){
 	if [[ "${osRelease}" == "debian" || "${osRelease}" == "ubuntu" ]]; then
-        if [[ "${osRelease}" == "debian" ]]; then
-            echo "deb http://deb.debian.org/debian buster-backports main contrib non-free" > /etc/apt/sources.list.d/buster-backports.list
-            echo "deb-src http://deb.debian.org/debian buster-backports main contrib non-free" >> /etc/apt/sources.list.d/buster-backports.list
-            ${sudoCmd} apt update
-        fi
+
+
+		if ! dpkg -l | grep -qw curl; then
+			${osSystemPackage} -y install wget curl git
+
+            if [[ "${osRelease}" == "debian" ]]; then
+                echo "deb http://deb.debian.org/debian buster-backports main contrib non-free" > /etc/apt/sources.list.d/buster-backports.list
+                echo "deb-src http://deb.debian.org/debian buster-backports main contrib non-free" >> /etc/apt/sources.list.d/buster-backports.list
+                ${sudoCmd} apt update -y
+            fi
+
+		fi
 
 		if ! dpkg -l | grep -qw wget; then
 			${osSystemPackage} -y install wget curl git
+
+            if [[ "${osRelease}" == "debian" ]]; then
+                echo "deb http://deb.debian.org/debian buster-backports main contrib non-free" > /etc/apt/sources.list.d/buster-backports.list
+                echo "deb-src http://deb.debian.org/debian buster-backports main contrib non-free" >> /etc/apt/sources.list.d/buster-backports.list
+                ${sudoCmd} apt update -y
+            fi
+
 		fi
 
         if ! dpkg -l | grep -qw bc; then
@@ -1677,7 +1691,7 @@ function installDebianUbuntuKernel(){
 
             echo "deb http://deb.debian.org/debian buster-backports main contrib non-free" > /etc/apt/sources.list.d/buster-backports.list
             echo "deb-src http://deb.debian.org/debian buster-backports main contrib non-free" >> /etc/apt/sources.list.d/buster-backports.list
-            ${sudoCmd} apt update
+            ${sudoCmd} apt update -y
 
             listAvailableLinuxKernel
             
@@ -2957,7 +2971,7 @@ function start_menu(){
 
     if [[ ${configLanguage} == "cn" ]] ; then
     green " =================================================="
-    green " Linux 内核 一键安装脚本 | 2022-1-27 | By jinwyp | 系统支持：centos7+ / debian10+ / ubuntu16.04+"
+    green " Linux 内核 一键安装脚本 | 2022-3-27 | By jinwyp | 系统支持：centos7+ / debian10+ / ubuntu16.04+"
     green " Linux 内核 4.9 以上都支持开启BBR, 如要开启BBR Plus 则需要安装支持BBR Plus的内核 "
     red " 在任何生产环境中请谨慎使用此脚本, 升级内核有风险, 请做好备份！在某些VPS会导致无法启动! "
     green " =================================================="
@@ -3038,7 +3052,7 @@ function start_menu(){
     else
 
     green " =================================================="
-    green " Linux kernel install script | 2022-1-27 | By jinwyp | OS support：centos7+ / debian10+ / ubuntu16.04+"
+    green " Linux kernel install script | 2022-3-27 | By jinwyp | OS support：centos7+ / debian10+ / ubuntu16.04+"
     green " Enable bbr require linux kernel higher than 4.9. Enable bbr plus require special bbr plus kernel "
     red " Please use this script with caution in production. Backup your data first! Upgrade linux kernel will cause VPS unable to boot sometimes."
     green " =================================================="
